pyproject.toml
[tool.poetry]
name = "PodcasterBot"
version = "0.1.0"
description = ""
authors = ["_Counter021_"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.15"
pafy = "^0.5.5"
youtube_dl = "^2021.6.6"
pydub = "^0.25.1"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
